requirements/development.txt
-r common.txt
django-debug-toolbar==3.2.2